release 5.4.0: festival days follow the reference almanac, Yoni and Brahma Muhurta corrected, DST year walks fixed, faster
Output-only minor release: nothing removed, renamed or retyped in either language. One new TS export
(computeVimshottariPratyantarIn); the Go module adds the Read* table readers and a few helpers.

- festivals: vriddhi and kala day rules checked against the reference almanac; new keys holika_dahan and
  karthigai_deepam; holi is the day after Holika Dahan; Onam once a year; Shravan Somvar follows masaSystem
- matching: published 4/3/2/1/0 Yoni table; Pathu Porutham Yoni is the Tamil enemy-pair test
- muhurta: Brahma Muhurta is the 14th night muhurta
- calendar: DST-correct year listings, Sankranti and the solar new year on the right day west of UTC,
  Odisha new year is Pana Sankranti, 1900/2100 and two-digit-year edges
- jyotish: birth-period dasha clipping, Nathonatha continuity, Sade Sati retrograde re-entry, Kemadruma
- inputs that hung or returned garbage now throw PanchangError codes
- performance: 1.4x on a default day up to 18x on Sade Sati, output byte-identical
- parity: g2 and g3 pins attributed byte for byte and re-pinned

+++ source/go/internal/astronomy/trig.go
@@ -14,7 +14,7 @@ const inversePI = 1 / jsnum.PI
 
 func reduce(x float64) (r float64, odd bool) {
 	q := jsnum.Round(x * inversePI)
-	r = ((x - q*pi1) - q*pi2) - q*pi3
+	r = float64(float64(x-float64(q*pi1))-float64(q*pi2)) - float64(q*pi3)
 	return r, int64(q)&1 == 1
 }
